[Security Solution] moving the alert description flyout component to shared package
