+++ src/platform/packages/shared/kbn-flyout-ui/moon.yml
@@ -28,6 +28,7 @@ tags:
 fileGroups:
   src:
     - '**/*.ts'
+    - '**/*.tsx'
     - '!target/**/*'
 tasks:
   jest:
