+++ src/platform/packages/shared/kbn-security-solution-common/index.ts
@@ -6,3 +6,5 @@
  * your election, the "Elastic License 2.0", the "GNU Affero General Public
  * License v3.0 only", or the "Server Side Public License, v 1".
  */
+
+export { FLYOUT_STORAGE_KEYS } from './constants/flyout_storage_keys';
