+++ src/platform/packages/shared/kbn-security-solution-common/moon.yml
@@ -28,6 +28,7 @@ tags:
 fileGroups:
   src:
     - '**/*.ts'
+    - '**/*.tsx'
     - '!target/**/*'
 tasks:
   jest:
