+++ src/platform/packages/shared/kbn-security-solution-flyout/moon.yml
@@ -17,7 +17,10 @@ project:
   owner: '@elastic/security-threat-hunting-investigations'
   metadata:
     sourceRoot: src/platform/packages/shared/kbn-security-solution-flyout
-dependsOn: []
+dependsOn:
+  - '@kbn/discover-utils'
+  - '@kbn/i18n'
+  - '@kbn/i18n-react'
 tags:
   - shared-browser
   - package
