+++ src/platform/packages/shared/kbn-security-solution-flyout/tsconfig.json
@@ -10,5 +10,5 @@
   "exclude": [
     "target/**/*"
   ],
-  "kbn_references": []
+  "kbn_references": ["@kbn/discover-utils", "@kbn/i18n", "@kbn/i18n-react"]
 }
