+++ src/platform/plugins/shared/discover/moon.yml
@@ -134,6 +134,9 @@ dependsOn:
   - '@kbn/es-query-constants'
   - '@kbn/esql'
   - '@kbn/controls-schemas'
+  - '@kbn/flyout-ui'
+  - '@kbn/security-solution-common'
+  - '@kbn/security-solution-flyout'
 tags:
   - plugin
   - prod
