+++ src/platform/plugins/shared/discover/public/context_awareness/profile_providers/security/components/test_ids.ts
@@ -1,0 +1,12 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the "Elastic License
+ * 2.0", the "GNU Affero General Public License v3.0 only", and the "Server Side
+ * Public License v 1"; you may not use this file except in compliance with, at
+ * your election, the "Elastic License 2.0", the "GNU Affero General Public
+ * License v3.0 only", or the "Server Side Public License, v 1".
+ */
+
+import { PREFIX } from '@kbn/security-solution-flyout';
+
+export const ABOUT_SECTION_TEST_ID = `${PREFIX}AboutSection` as const;
