+++ src/platform/plugins/shared/discover/public/context_awareness/profile_providers/security/translations.ts
@@ -9,12 +9,6 @@
 
 import { i18n } from '@kbn/i18n';
 
-export const exploreRowActionLabel = (isAlert: boolean) =>
-  i18n.translate('discover.profile.security.rowAction.exploreButtonLabel', {
-    values: { isAlert },
-    defaultMessage: 'Explore {isAlert, select, true {Alert} other {Event}} in Security',
-  });
-
 export const overviewTabTitle = (isAlert: boolean) =>
   i18n.translate('discover.profile.security.flyout.overviewTabTitle', {
     values: { isAlert },
