+++ x-pack/solutions/security/plugins/security_solution/moon.yml
@@ -275,6 +275,9 @@ dependsOn:
   - '@kbn/babel-register'
   - '@kbn/core-security-common'
   - '@kbn/entity-store'
+  - '@kbn/flyout-ui'
+  - '@kbn/security-solution-common'
+  - '@kbn/security-solution-flyout'
 tags:
   - plugin
   - prod
