+++ x-pack/solutions/security/plugins/security_solution/public/flyout/document_details/right/components/insights_section.test.tsx
@@ -6,7 +6,7 @@
  */
 
 import React from 'react';
-import { render } from '@testing-library/react';
+import { act, render } from '@testing-library/react';
 import { DocumentDetailsContext } from '../../shared/context';
 import {
   CORRELATIONS_TEST_ID,
@@ -27,7 +27,7 @@ import { mockDataFormattedForFieldBrowser } from '../../shared/mocks/mock_data_f
 import { InsightsSection } from './insights_section';
 import { useAlertPrevalence } from '../../shared/hooks/use_alert_prevalence';
 import { useRiskScore } from '../../../../entity_analytics/api/hooks/use_risk_score';
-import { useExpandSection } from '../../../shared/hooks/use_expand_section';
+import { useExpandSection } from '@kbn/flyout-ui';
 import { useIsExperimentalFeatureEnabled } from '../../../../common/hooks/use_experimental_features';
 import { useSecurityDefaultPatterns } from '../../../../data_view_manager/hooks/use_security_default_patterns';
 
@@ -64,7 +64,14 @@ const from = '2022-04-05T12:00:00.000Z';
 const to = '2022-04-08T12:00:00.;000Z';
 const selectedPatterns = 'alerts';
 
-jest.mock('../../../shared/hooks/use_expand_section');
+jest.mock('@kbn/flyout-ui', () => {
+  const actual = jest.requireActual('@kbn/flyout-ui');
+  return {
+    ...actual,
+    useExpandSection: jest.fn(),
+  };
+});
+
 const mockUseGlobalTime = jest.fn().mockReturnValue({ from, to });
 jest.mock('../../../../common/containers/use_global_time', () => {
   return {
@@ -105,7 +112,11 @@ const renderInsightsSection = (contextValue: DocumentDetailsContext) =>
   );
 
 describe('<InsightsSection />', () => {
+  const mockUseExpandSection = jest.mocked(useExpandSection);
+
   beforeEach(() => {
+    jest.clearAllMocks();
+    mockUseExpandSection.mockReturnValue(true);
     (useIsExperimentalFeatureEnabled as jest.Mock).mockReturnValue(true);
     (useSecurityDefaultPatterns as jest.Mock).mockReturnValue({
       indexPatterns: ['index'],
@@ -126,7 +137,7 @@ describe('<InsightsSection />', () => {
     });
   });
 
-  it('should render insights component', () => {
+  it('should render insights component', async () => {
     const contextValue = {
       eventId: 'some_Id',
       getFieldsData: mockGetFieldsData,
@@ -134,26 +145,14 @@ describe('<InsightsSection />', () => {
 
     const wrapper = renderInsightsSection(contextValue);
 
-    expect(wrapper.getByTestId(INSIGHTS_HEADER_TEST_ID)).toBeInTheDocument();
-    expect(wrapper.getByTestId(INSIGHTS_HEADER_TEST_ID)).toHaveTextContent('Insights');
-    expect(wrapper.getByTestId(INSIGHTS_CONTENT_TEST_ID)).toBeInTheDocument();
+    await act(async () => {
+      expect(wrapper.getByTestId(INSIGHTS_HEADER_TEST_ID)).toHaveTextContent('Insights');
+      expect(wrapper.getByTestId(INSIGHTS_CONTENT_TEST_ID)).toBeInTheDocument();
+    });
   });
 
-  it('should render the component collapsed if value is false in local storage', () => {
-    (useExpandSection as jest.Mock).mockReturnValue(false);
-
-    const contextValue = {
-      eventId: 'some_Id',
-      dataFormattedForFieldBrowser: mockDataFormattedForFieldBrowser,
-      getFieldsData: mockGetFieldsData,
-    } as unknown as DocumentDetailsContext;
-
-    const wrapper = renderInsightsSection(contextValue);
-    expect(wrapper.getByTestId(INSIGHTS_CONTENT_TEST_ID)).not.toBeVisible();
-  });
-
-  it('should render the component expanded if value is true in local storage', () => {
-    (useExpandSection as jest.Mock).mockReturnValue(true);
+  it('should render the component collapsed if value is false in local storage', async () => {
+    mockUseExpandSection.mockReturnValue(false);
 
     const contextValue = {
       eventId: 'some_Id',
@@ -162,12 +161,27 @@ describe('<InsightsSection />', () => {
     } as unknown as DocumentDetailsContext;
 
     const wrapper = renderInsightsSection(contextValue);
-    expect(wrapper.getByTestId(INSIGHTS_CONTENT_TEST_ID)).toBeVisible();
+
+    await act(async () => {
+      expect(wrapper.getByTestId(INSIGHTS_CONTENT_TEST_ID)).not.toBeVisible();
+    });
   });
 
-  it('should render all children when event kind is signal', () => {
-    (useExpandSection as jest.Mock).mockReturnValue(true);
+  it('should render the component expanded if value is true in local storage', async () => {
+    const contextValue = {
+      eventId: 'some_Id',
+      dataFormattedForFieldBrowser: mockDataFormattedForFieldBrowser,
+      getFieldsData: mockGetFieldsData,
+    } as unknown as DocumentDetailsContext;
 
+    const wrapper = renderInsightsSection(contextValue);
+
+    await act(async () => {
+      expect(wrapper.getByTestId(INSIGHTS_CONTENT_TEST_ID)).toBeVisible();
+    });
+  });
+
+  it('should render all children when event kind is signal', async () => {
     const getFieldsData = (field: string) => {
       switch (field) {
         case 'event.kind':
@@ -182,15 +196,15 @@ describe('<InsightsSection />', () => {
 
     const { getByTestId } = renderInsightsSection(contextValue);
 
-    expect(getByTestId(`${INSIGHTS_ENTITIES_TEST_ID}LeftSection`)).toBeInTheDocument();
-    expect(getByTestId(`${INSIGHTS_THREAT_INTELLIGENCE_TEST_ID}LeftSection`)).toBeInTheDocument();
-    expect(getByTestId(`${CORRELATIONS_TEST_ID}LeftSection`)).toBeInTheDocument();
-    expect(getByTestId(`${PREVALENCE_TEST_ID}LeftSection`)).toBeInTheDocument();
+    await act(async () => {
+      expect(getByTestId(`${INSIGHTS_ENTITIES_TEST_ID}LeftSection`)).toBeInTheDocument();
+      expect(getByTestId(`${INSIGHTS_THREAT_INTELLIGENCE_TEST_ID}LeftSection`)).toBeInTheDocument();
+      expect(getByTestId(`${CORRELATIONS_TEST_ID}LeftSection`)).toBeInTheDocument();
+      expect(getByTestId(`${PREVALENCE_TEST_ID}LeftSection`)).toBeInTheDocument();
+    });
   });
 
-  it('should not render threat intel and correlations insights component when document is not signal', () => {
-    (useExpandSection as jest.Mock).mockReturnValue(true);
-
+  it('should not render threat intel and correlations insights component when document is not signal', async () => {
     const getFieldsData = (field: string) => {
       switch (field) {
         case 'event.kind':
@@ -205,11 +219,13 @@ describe('<InsightsSection />', () => {
 
     const { getByTestId, queryByTestId } = renderInsightsSection(contextValue);
 
-    expect(getByTestId(`${INSIGHTS_ENTITIES_TEST_ID}LeftSection`)).toBeInTheDocument();
-    expect(
-      queryByTestId(`${INSIGHTS_THREAT_INTELLIGENCE_TEST_ID}LeftSection`)
-    ).not.toBeInTheDocument();
-    expect(getByTestId(`${CORRELATIONS_TEST_ID}LeftSection`)).toBeInTheDocument();
-    expect(getByTestId(`${PREVALENCE_TEST_ID}LeftSection`)).toBeInTheDocument();
+    await act(async () => {
+      expect(getByTestId(`${INSIGHTS_ENTITIES_TEST_ID}LeftSection`)).toBeInTheDocument();
+      expect(
+        queryByTestId(`${INSIGHTS_THREAT_INTELLIGENCE_TEST_ID}LeftSection`)
+      ).not.toBeInTheDocument();
+      expect(getByTestId(`${CORRELATIONS_TEST_ID}LeftSection`)).toBeInTheDocument();
+      expect(getByTestId(`${PREVALENCE_TEST_ID}LeftSection`)).toBeInTheDocument();
+    });
   });
 });
