+++ x-pack/solutions/security/plugins/security_solution/public/flyout/document_details/right/components/response_section.tsx
@@ -9,9 +9,8 @@ import React, { memo, useMemo } from 'react';
 import { FormattedMessage } from '@kbn/i18n-react';
 import { EuiCallOut } from '@elastic/eui';
 import { i18n } from '@kbn/i18n';
-import { useExpandSection } from '../../../shared/hooks/use_expand_section';
+import { ExpandableSection, useExpandSection } from '@kbn/flyout-ui';
 import { ResponseButton } from './response_button';
-import { ExpandableSection } from '../../../shared/components/expandable_section';
 import { useDocumentDetailsContext } from '../../shared/context';
 import { getField } from '../../shared/utils';
 import { EventKind } from '../../shared/constants/event_kinds';
