+++ x-pack/solutions/security/plugins/security_solution/public/flyout/document_details/right/components/visualizations_section.tsx
@@ -8,10 +8,9 @@
 import React, { memo } from 'react';
 import { EuiSpacer } from '@elastic/eui';
 import { FormattedMessage } from '@kbn/i18n-react';
-import { useExpandSection } from '../../../shared/hooks/use_expand_section';
+import { ExpandableSection, useExpandSection } from '@kbn/flyout-ui';
 import { AnalyzerPreviewContainer } from './analyzer_preview_container';
 import { SessionPreviewContainer } from './session_preview_container';
-import { ExpandableSection } from '../../../shared/components/expandable_section';
 import { VISUALIZATIONS_TEST_ID } from './test_ids';
 import { GraphPreviewContainer } from './graph_preview_container';
 import { useDocumentDetailsContext } from '../../shared/context';
