+++ x-pack/solutions/security/plugins/security_solution/public/flyout/document_details/shared/constants/local_storage.ts
@@ -5,9 +5,10 @@
  * 2.0.
  */
 
+import { FLYOUT_STORAGE_KEYS as COMMON_FLYOUT_STORAGE_KEYS } from '@kbn/security-solution-common';
+
 export const FLYOUT_STORAGE_KEYS = {
-  OVERVIEW_TAB_EXPANDED_SECTIONS:
-    'securitySolution.documentDetailsFlyout.overviewSectionExpanded.v8.14',
+  ...COMMON_FLYOUT_STORAGE_KEYS,
   RIGHT_PANEL_SELECTED_TABS: 'securitySolution.documentDetailsFlyout.rightPanel.selectedTabs.v8.14',
   TABLE_TAB_STATE: 'securitySolution.documentDetailsFlyout.tableTabState.v8.19',
   TABLE_TAB_TOUR: 'securitySolution.documentDetailsFlyout.tableTabTourState.v8.19',
